Fix CA1028
See https://docs.microsoft.com/en-us/visualstudio/code-quality/ca1028-enum-storage-should-be-int32?view=vs-2017

+++ src/DotNet/CallingConvention.cs
@@ -7,7 +7,7 @@ namespace dnlib.DotNet {
 	/// See CorHdr.h/CorCallingConvention
 	/// </summary>
 	[Flags]
-	public enum CallingConvention : byte {
+	public enum CallingConvention : int {
 		/// <summary>The managed calling convention</summary>
 		Default			= 0x0,
 		/// <summary/>
