+++ src/DotNet/Emit/Code.cs
@@ -5,7 +5,7 @@
 	/// A CIL opcode. If the high byte is 0 or if it's <see cref="UNKNOWN1"/>, it's a 1-byte opcode,
 	/// else it's a two-byte opcode and the highest byte is the first byte of the opcode.
 	/// </summary>
-	public enum Code : ushort {
+	public enum Code : int {
 #pragma warning disable 1591	// disable XML doc warning
 		UNKNOWN1		= 0x0100,
 		UNKNOWN2		= 0x0101,
