+++ src/DotNet/Emit/OperandType.cs
@@ -6,7 +6,7 @@ namespace dnlib.DotNet.Emit {
 	/// <summary>
 	/// CIL opcode operand type
 	/// </summary>
-	public enum OperandType : byte {
+	public enum OperandType : int {
 		/// <summary>4-byte relative instruction offset</summary>
 		InlineBrTarget,
 		/// <summary>4-byte field token (<see cref="Table.Field"/> or <see cref="Table.MemberRef"/>)</summary>
