+++ src/DotNet/MD/ColumnSize.cs
@@ -4,7 +4,7 @@
 	/// <summary>
 	/// MD table column size
 	/// </summary>
-	public enum ColumnSize : byte {
+	public enum ColumnSize : int {
 		/// <summary>RID into Module table</summary>
 		Module,
 		/// <summary>RID into TypeRef table</summary>
